requirements.txt
Django==3.0.5
djangorestframework==3.11.0
geopy==1.21.0
django-redis==4.11.0
django-easy-timezones-redux==1.0.4
psycopg2-binary==2.8.5
gunicorn==20.0.4
celery==4.4.2
